add config files to build wheel for easier use in a headless environment
