+++ pyproject.toml
@@ -1,0 +1,4 @@
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
+
