+++ setup.cfg
@@ -1,0 +1,10 @@
+[metadata]
+name = sigkit
+version = 1.2.1
+license = "MIT"
+long_description = file: README.md
+
+[options]
+install_requires = flatbuffers
+packages=find:
+
