+++ sigkit/FlatbufSignatureLibrary/Pattern.py
