+++ sigkit/FlatbufSignatureLibrary/TrieNode.py
