+++ sigkit/compute_sig.py
