+++ sigkit/sigexplorer.py
