+++ sigkit/signaturelibrary.py
