+++ sigkit/trie_ops.py
